Conflicts:
- gRPC.podspec
  - Only had non-trivial changes in the core file list, which will need to be
    regenerated (in gRPC-Core.podspec).
- src/objective-c/BoringSSL.podspec
  - Had trivial conflicts in the version.
- src/objective-c/examples/RemoteTestClient/RemoteTest.podspec
  - Trivial conflicts in quoting.
- src/objective-c/examples/Sample/Sample.xcodeproj/project.pbxproj and
src/objective-c/examples/SwiftSample/SwiftSample.xcodeproj/project.pbxproj
  - The master version is used, pending testing. The 0.14 version had emoji and
    some unneeded entries.
- src/objective-c/tests/Podfile
  - Added CronetFramework pod, and warning silencing from master.
- templates/gRPC.podspec.template
  - Deleted.
- third_party/protobuf
  - Using master commit, but need to verify if it works for frameworks.

+++ src/core/lib/support/log_linux.c
@@ -95,9 +95,9 @@ void gpr_default_log(gpr_log_func_args *args) {
     strcpy(time_buffer, "error:strftime");
   }
 
-  gpr_asprintf(&prefix, "%s%s.%09d %7tu %s:%d]",
+  gpr_asprintf(&prefix, "%s%s.%09" PRId32 " %7ld %s:%d]",
                gpr_log_severity_string(args->severity), time_buffer,
-               (int)(now.tv_nsec), gettid(), display_file, args->line);
+               now.tv_nsec, gettid(), display_file, args->line);
 
   fprintf(stderr, "%-60s %s\n", prefix, args->message);
   gpr_free(prefix);
